backend/requirements.txt
# ==========================
# Django and DRF
# ==========================
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3

# ==========================
# Database
# ==========================
psycopg2-binary==2.9.9

# ==========================
# Channels (WebSocket)
# ==========================
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# ==========================
# Celery (Async Tasks)
# ==========================
celery==5.4.0
redis==5.0.8
django-celery-beat==2.7.0
django-celery-results==2.5.1  # ✅ добавить для хранения результатов задач
flower==2.0.1

# ==========================
# API Documentation
# ==========================
drf-spectacular==0.27.2

# ==========================
# Files and Reports
# ==========================
Pillow==10.4.0
django-import-export==4.1.1
weasyprint==62.3
openpyxl==3.1.5
xlsxwriter==3.2.0

# ==========================
# Environment Variables
# ==========================
python-dotenv==1.0.1

# ==========================
# Security
# ==========================
django-ratelimit==4.1.0

# ==========================
# Utilities
# ==========================
python-dateutil==2.9.0
pytz==2024.1
requests==2.32.3

# ==========================
# Testing
# ==========================
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.0  # ✅ удобно для тестовых данных

# ==========================
# Code Formatting (Dev only)
# ==========================
black==24.4.2
isort==5.13.2
flake8==7.1.1